Move reasoning stuff to reasoning.js
